PDM/server/requirements.txt
Django==1.11.1
celery==4.0.2
django-rest-swagger==2.1.2
djangorestframework>=3.7
gunicorn==19.7.1
presslabs-sugar==1.1.0
raven==6.0.0
redis==2.10.5
requests==2.13.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
